requirements.txt
python-telegram-bot==20.7
anthropic
llama-index==0.14.18
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb
PyMuPDF
gitpython
sentence-transformers
python-dotenv